refactor-#35: rename db module to database module
