+++ src/app.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { CqrsModule } from '@nestjs/cqrs';
 
-import { DbModule } from './db/db.module';
+import { DatabaseModule } from './database/database.module';
 import { CatalogModule } from './catalog/catalog.module';
 import { HealthModule } from './health/health.module';
 import { UtilsModule } from './utils/utils.module';
@@ -13,7 +13,7 @@ import { EsModule } from './es/es.module';
     ConfigModule.forRoot(),
     CqrsModule,
     EsModule,
-    DbModule,
+    DatabaseModule,
     CatalogModule,
     HealthModule,
     UtilsModule,
