+++ src/catalog/cars/queries/handlers/get-all-cars.handler.ts
@@ -2,9 +2,9 @@ import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
-import { CarsRepository } from '../../../../db/repositories';
+import { CarsRepository } from '../../../../database/repositories';
 import { GetAllCarsQuery } from '../impl/get-all-cars.query';
-import { CarsEntity } from '../../../../db/entities/cars.entity';
+import { CarsEntity } from '../../../../database/entities/cars.entity';
 
 @QueryHandler(GetAllCarsQuery)
 export class GetAllCarsHandler implements IQueryHandler<GetAllCarsQuery> {
