+++ src/catalog/cars/services/cars.service.ts
@@ -9,7 +9,7 @@ import {
 import { CreateCarsDto, UpdateCarsDto } from '../dto';
 import { TextResponseModel } from '../models/text-response.model';
 import { GetSingleCarQuery, GetAllCarsQuery } from '../queries/impl';
-import { CarsEntity } from '../../../db/entities/cars.entity';
+import { CarsEntity } from '../../../database/entities/cars.entity';
 
 @Injectable()
 export class CarsService {
