+++ src/catalog/flight/commands/handlers/update-flight.handler.ts
@@ -2,9 +2,9 @@ import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
-import { FlightRepository } from '../../../../db/repositories';
+import { FlightRepository } from '../../../../database/repositories';
 import { UpdateFlightCommand } from '../impl';
-import { FlightEntity } from '../../../../db/entities/flight.entity';
+import { FlightEntity } from '../../../../database/entities/flight.entity';
 
 @CommandHandler(UpdateFlightCommand)
 export class UpdateFlightHandler
