+++ src/catalog/flight/controllers/flight.controller.ts
@@ -2,7 +2,7 @@ import { Controller } from '@nestjs/common';
 import { MessagePattern } from '@nestjs/microservices';
 
 import { CreateFlightDto, UpdateFlightDto } from '../dto';
-import { FlightEntity } from '../../../db/entities/flight.entity';
+import { FlightEntity } from '../../../database/entities/flight.entity';
 import { FlightService } from '../services/flight.service';
 import { TextResponseModel } from '../models/text-response.model';
