+++ src/catalog/flight/services/flight.service.ts
@@ -9,7 +9,7 @@ import {
 import { CreateFlightDto, UpdateFlightDto } from '../dto';
 import { TextResponseModel } from '../models/text-response.model';
 import { GetSingleFlightQuery, GetAllFlightsQuery } from '../queries/impl';
-import { FlightEntity } from '../../../db/entities/flight.entity';
+import { FlightEntity } from '../../../database/entities/flight.entity';
 
 @Injectable()
 export class FlightService {
