+++ src/catalog/hotel/controllers/hotel.controller.ts
@@ -2,7 +2,7 @@ import { Controller } from '@nestjs/common';
 import { MessagePattern } from '@nestjs/microservices';
 
 import { CreateHotelDto, UpdateHotelDto } from '../dto';
-import { HotelEntity } from '../../../db/entities/hotel.entity';
+import { HotelEntity } from '../../../database/entities/hotel.entity';
 import { HotelService } from '../services/hotel.service';
 import { TextResponseModel } from '../models/text-response.model';
