+++ src/database/entities/flight.entity.ts
