+++ src/database/entities/hotel.entity.ts
