+++ src/database/subscribers/hotel.subscriber.ts
